[Perf Improver] count_runs_by_status() in health metrics — fix 200-run cap + N→4 rows (#329)
* perf: count_runs_by_status() in health metrics — fix 200-run cap + N→4 rows

load_metrics_snapshot() previously called list_runs(None, 200) to count
runs by status in Rust. This had two problems:

1. Correctness: status counts silently capped at 200 — any runs beyond
   the first 200 are missed in the metrics response.

2. Performance: fetches up to 200 full OrchestrationRun rows (all
   columns, full deserialization) just to count by status field.

Add OrchestrationStore::count_runs_by_status() backed by:

  SELECT status, COUNT(*) FROM orchestration_runs GROUP BY status

This returns at most 4 rows regardless of table size and covers all
runs (no cap). Replace list_runs(None, 200) in load_metrics_snapshot()
with count_runs_by_status().

Benchmark results (in-memory WAL):
  10 rows:  list_runs_200  12.8µs → count_by_status  2.1µs  (~6× faster)
  50 rows:  list_runs_200  55.0µs → count_by_status  4.4µs  (~12× faster)
  200 rows: list_runs_200 200µs  → count_by_status  13.3µs  (~15× faster)

The speedup grows with run count; beyond 200 runs count_by_status is
the only option that gives correct results.

Also:
- Add 2 unit tests for count_runs_by_status (empty DB, mixed statuses)
- Add or_count_vs_list_200 benchmark group to orchestration_store bench
- Update RunMetrics serialization fields from usize to i64 (matching
  the i64 returned by SQLite COUNT)

293 tests pass.


* style: fix formatting in benchmark and mod.rs import order


---------

+++ crates/opengoose-web/src/routes/health/responses.rs
@@ -25,10 +25,10 @@ pub(crate) struct QueueMetrics {
 
 #[derive(serde::Serialize)]
 pub(crate) struct RunMetrics {
-    pub(crate) running: usize,
-    pub(crate) completed: usize,
-    pub(crate) failed: usize,
-    pub(crate) suspended: usize,
+    pub(crate) running: i64,
+    pub(crate) completed: i64,
+    pub(crate) failed: i64,
+    pub(crate) suspended: i64,
 }
 
 #[derive(serde::Serialize)]
